pipeline/assets/cdk-macro/requirements.txt
boto3
cfn-flip
aws-cdk-lib
